+++ mitra_validators/src/polls.rs
@@ -1,10 +1,25 @@
 use mitra_models::polls::types::PollData;
+use mitra_utils::html::clean_html_all;
 
 use super::errors::ValidationError;
 
 const POLL_OPTION_COUNT_MIN: usize = 2;
 pub const POLL_OPTION_COUNT_MAX: usize = 10;
 pub const POLL_OPTION_NAME_LENGTH_MAX: usize = 1000;
+
+fn clean_poll_option_name(name: &str) -> String {
+    clean_html_all(name)
+}
+
+fn validate_poll_option_name(option_name: &str) -> Result<(), ValidationError> {
+    if option_name.len() > POLL_OPTION_NAME_LENGTH_MAX {
+        return Err(ValidationError("poll option name is too long"));
+    };
+    if option_name != clean_poll_option_name(option_name) {
+        return Err(ValidationError("option name has not been sanitized"));
+    };
+    Ok(())
+}
 
 pub fn validate_poll_data(poll_data: &PollData) -> Result<(), ValidationError> {
     if poll_data.results.len() < POLL_OPTION_COUNT_MIN {
@@ -15,9 +30,7 @@ pub fn validate_poll_data(poll_data: &PollData) -> Result<(), ValidationError> {
     };
     let mut unique_options = vec![];
     for result in &poll_data.results {
-        if result.option_name.len() > POLL_OPTION_NAME_LENGTH_MAX {
-            return Err(ValidationError("poll option name is too long"));
-        };
+        validate_poll_option_name(&result.option_name)?;
         if !unique_options.contains(&&result.option_name) {
             unique_options.push(&result.option_name);
         } else {
@@ -31,6 +44,12 @@ pub fn validate_poll_data(poll_data: &PollData) -> Result<(), ValidationError> {
 mod tests {
     use mitra_models::polls::types::PollResult;
     use super::*;
+
+    #[test]
+    fn test_validate_poll_option_name_sanitized() {
+        let name = "test <span>html</span>";
+        assert_eq!(validate_poll_option_name(&name).is_ok(), false);
+    }
 
     #[test]
     fn test_poll_data_unique_options() {