+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Changed MSRV to 1.68.2.
 - Write warning to log if actor and actor key have different origins.
 - Reject portable actors without WebFinger address.
+- Truncate long display names instead of rejecting actor.
 
 ## Fixed